renaming of power_spectra_generator
