+++ flip/power_spectra/__init__.py
@@ -1,3 +1,4 @@
 """Init file of the flip.power_spectra package."""
+
 from . import class_engine, models
-from .power_spectra_generator import compute_power_spectra
+from .generator import compute_power_spectra
